+++ src/builders/encryption.ts
@@ -1,6 +1,8 @@
 import { evaluate } from './ValidationError';
 import { amount } from '../primitives';
 import {
+  AccessCondition,
+  AccessConditionSchema,
   AndCondition,
   AndConditionSchema,
   AnyCondition,
@@ -21,6 +23,7 @@ import {
   OrConditionSchema,
   ProfileOwnershipCondition,
   ProfileOwnershipConditionSchema,
+  SimpleCondition,
   refineNftOwnershipCondition,
 } from '../publication';
 
@@ -223,7 +226,7 @@ export function followCondition(input: FollowConditionDetails): FollowCondition 
  *
  * @category Helpers
  */
-export function andCondition<T extends AnyCondition[]>(criteria: T): AndCondition<T[number]> {
+export function andCondition<T extends SimpleCondition[]>(criteria: T): AndCondition<T[number]> {
   return evaluate(
     AndConditionSchema.safeParse({
       type: ConditionType.AND,
@@ -237,9 +240,23 @@ export function andCondition<T extends AnyCondition[]>(criteria: T): AndConditio
  *
  * @category Helpers
  */
-export function orCondition<T extends AnyCondition[]>(criteria: T): OrCondition<T[number]> {
+export function orCondition<T extends SimpleCondition[]>(criteria: T): OrCondition<T[number]> {
   return evaluate(
     OrConditionSchema.safeParse({
+      type: ConditionType.OR,
+      criteria,
+    }),
+  );
+}
+
+/**
+ * Creates the access condition specification for a given publication.
+ *
+ * @internal
+ */
+export function accessCondition(criteria: AnyCondition[]): AccessCondition {
+  return evaluate(
+    AccessConditionSchema.safeParse({
       type: ConditionType.OR,
       criteria,
     }),